changed gmail_id + thread_id in topics + emails from Long to String

+++ src/main/scala/com/textMailer/IO/actors/EmailActor.scala
@@ -27,7 +27,7 @@ class EmailActor extends Actor {
         uid <- userId
         tid <- threadId
       } yield(uid, tid)) match {
-        case Some(ids) => EmailTopicIO().asyncFind(List(Eq("user_id", ids._1), Eq("thread_id", ids._2.toLong)), 40)
+        case Some(ids) => EmailTopicIO().asyncFind(List(Eq("user_id", ids._1), Eq("thread_id", ids._2)), 40)
         case None => Future(List())
       }
